backend/requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.2
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.0
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Supabase
supabase==2.3.0
gotrue==2.1.0
storage3==0.7.0

# Auth / Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
cryptography==41.0.7
PyJWT==2.8.0

# Redis / Celery
celery==5.3.6
redis==5.0.1
celery[redis]==5.3.6

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Email
resend==0.7.0

# Rate Limiting
slowapi==0.1.9

# Logging
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.23.2
pytest-cov==4.1.0
httpx==0.25.2
factory-boy==3.3.1

# Utils
python-slugify==8.0.1
python-dotenv==1.0.0
orjson==3.9.10
